+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_enclave_functions.go
@@ -500,6 +500,8 @@ func (backend *DockerKurtosisBackend) getEnclaveStatusAndContainers(
 		}
 		if isContainerRunning {
 			resultEnclaveStatus = enclave.EnclaveStatus_Running
+			//Enclave is considered running if we found at least one container running
+			break
 		}
 	}